Remove serialize_spec and rename build_spec to build_model
