+++ bioimageio/core/build_spec/__init__.py
@@ -1,1 +1,1 @@
-from .build_spec import add_weights, build_spec, serialize_spec
+from .build_model import add_weights, build_model
